requirements copy.txt
schema_entry
colorama
gevent
wheel
pep8
coverage
mypy-extensions
pydocstyle
autopep8
sphinx
recommonmark
sphinxcontrib-napoleon
sphinx-intl
mypy
lxml
mypy-extensions
boom
chardet
grpclib
pyyaml
colorama
termcolor
requests_oauthlib
sphinx_rtd_theme